Time: 0 ms (100.00%) | Memory: 8.7 MB (13.35%) - LeetSync

+++ 69-sqrtx/sqrtx.cpp
@@ -10,10 +10,10 @@ public:
         while (l <= r) {
             int mid = l + (r - l) / 2;
 
-            if ((long)mid * mid == x)
+            if ((long long)mid * mid == x)
                 return mid;
 
-            else if ((long)mid * mid < x) {
+            else if ((long long)mid * mid < x) {
                 ans = mid;
                 l = mid + 1;
             }
